dom/base/test/chrome/test_registerElement_ep.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1130028
-->
<window title="Mozilla Bug 1130028"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1130028"
     target="_blank">Mozilla Bug 1130028</a>
  <iframe onload="startTests()" id="frame" src="http://example.com/chrome/dom/base/test/chrome/frame_registerElement_content.html"></iframe>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

  Components.utils.import("resource://gre/modules/Services.jsm");

  /** Test for Bug 1130028 **/
  SimpleTest.waitForExplicitFinish();

  function finishTest(canSeePrototype) {
    ok(true, "createdCallback called when reigsterElement was called with an extended principal.");
    ok(canSeePrototype, "createdCallback should be able to see custom prototype.");
    SimpleTest.finish();
  }

  function startTests() {
    var frame = $("frame");

    // Create a sandbox with an extended principal then run a script that registers a custom element in the sandbox.
    var sandbox = Components.utils.Sandbox([frame.contentWindow], { sandboxPrototype: frame.contentWindow });
    sandbox.finishTest = finishTest;
    Services.scriptloader.loadSubScript("chrome://mochitests/content/chrome/dom/base/test/chrome/registerElement_ep.js", sandbox);
  }

  ]]></script>
</window>
